Unit tests and client side edit form changes
- Client side uses checkboxes for role selection now.
- Tested api's for save and edit
- Possible bug found regarding default email_confirmed_at date in user entity unit tests:
UserEntity passes an arg for email verification time as “server_default='0000-00-00 00:00:00'” which
causes a “OverflowError: date value out of range” in the tzinfo.py library when running the unit tests.
This error is either not present or does nothing in production. Further changing the date to a recent
one sidesteps the problem